pyproject.toml
[project]
name = "captain-claw"
version = "0.1.0"
description = "A powerful console-based AI agent"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Stevica Kuharski", email = "kstevica@gmail.com" }
]
keywords = ["ai", "agent", "cli", "terminal", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "litellm>=1.40.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "aiosqlite>=0.19.0",
    "tiktoken>=0.5.0",
    "httpx>=0.25.0",
    "aiofiles>=23.0.0",
    "structlog>=23.0.0",
    "pyyaml>=6.0.0",
    "rich>=13.0.0",
    "blessed>=1.20.0",
    "python-dotenv>=1.0.0",
    "glob2>=0.7",
    "tenacity>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
vector = [
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
]

[project.scripts]
captain-claw = "captain_claw.main:main"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
